Move getImageSpritePixels to ImageUtil
