+++ runelite-client/src/main/java/net/runelite/client/plugins/interfacestyles/InterfaceStylesPlugin.java
@@ -29,7 +29,6 @@ package net.runelite.client.plugins.interfacestyles;
 import com.google.common.eventbus.Subscribe;
 import com.google.inject.Provides;
 import java.awt.image.BufferedImage;
-import java.awt.image.PixelGrabber;
 import java.io.IOException;
 import java.io.InputStream;
 import javax.imageio.ImageIO;
@@ -47,6 +46,7 @@ import net.runelite.client.config.ConfigManager;
 import net.runelite.client.game.SpriteManager;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
+import net.runelite.client.util.ImageUtil;
 
 @Slf4j
 @PluginDescriptor(
@@ -195,7 +195,7 @@ public class InterfaceStylesPlugin extends Plugin
 		{
 			log.debug("Loading: " + filePath);
 			BufferedImage spriteImage = ImageIO.read(inputStream);
-			return getImageSpritePixels(spriteImage, client);
+			return ImageUtil.getImageSpritePixels(spriteImage, client);
 		}
 		catch (IOException ex)
 		{
@@ -207,23 +207,6 @@ public class InterfaceStylesPlugin extends Plugin
 		}
 
 		return null;
-	}
-
-	public static SpritePixels getImageSpritePixels(BufferedImage image, Client client)
-	{
-		int[] pixels = new int[image.getWidth() * image.getHeight()];
-
-		try
-		{
-			new PixelGrabber(image, 0, 0, image.getWidth(), image.getHeight(), pixels, 0, image.getWidth())
-				.grabPixels();
-		}
-		catch (InterruptedException ex)
-		{
-			log.debug("PixelGrabber was interrupted: ", ex);
-		}
-
-		return client.createSpritePixels(pixels, image.getWidth(), image.getHeight());
 	}
 
 	private void adjustWidgetDimensions()
@@ -284,7 +267,7 @@ public class InterfaceStylesPlugin extends Plugin
 
 		if (compassImage != null)
 		{
-			SpritePixels compass = getImageSpritePixels(compassImage, client);
+			SpritePixels compass = ImageUtil.getImageSpritePixels(compassImage, client);
 			client.setCompass(compass);
 		}
 	}
